requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
gunicorn==20.1.0
pandas
pyarrow
Werkzeug<3.0.0
